+++ Samples/SentrySampleShared/SentrySampleShared/SentrySDKWrapper.swift
@@ -182,6 +182,10 @@ public struct SentrySDKWrapper {
         options.enableFileManagerSwizzling = !SentrySDKOverrides.Other.disableFileManagerSwizzling.boolValue
         options.experimental.enableUnhandledCPPExceptionsV2 = true
         options.experimental.enableStandaloneAppStartTracing = true
+
+#if os(macOS) && !SENTRY_NO_UI_FRAMEWORK
+        options.enableUncaughtNSExceptionReporting = true
+#endif
     }
 
     func configureInitialScope(scope: Scope, options: Options) -> Scope {